requirements-dev.txt
numpy==2.4.4
pillow==12.2.0
pytest>=8.0
pygame>=2.5
